Refactor Milkdrop runtime and compiler seams
