+++ assets/js/milkdrop/runtime.ts
@@ -1,13 +1,14 @@
+export {
+  applyMilkdropInteractionResponse,
+  buildMilkdropInputSignalOverrides,
+  getMilkdropDetailScale,
+} from './runtime/interaction-response';
+
 import {
   clearDebugSnapshot,
   isAgentMode,
   setDebugSnapshot,
 } from '../core/agent-api.ts';
-import type { ShaderQuality } from '../core/performance-panel';
-import {
-  isCompatibilityModeEnabled,
-  setCompatibilityMode,
-} from '../core/render-preferences';
 import { getCachedRendererCapabilities } from '../core/renderer-capabilities.ts';
 import {
   type AdaptiveQualityController,
@@ -20,7 +21,6 @@ import {
 } from '../core/settings-panel.ts';
 import type { ToyRuntimeFrame, ToyRuntimeInstance } from '../core/toy-runtime';
 import type { QualityPresetManager } from '../utils/toy-settings';
-import type { UnifiedInputState } from '../utils/unified-input';
 import { createMilkdropCatalogStore } from './catalog-store';
 import { consumeRequestedMilkdropCollectionSelection } from './collection-intent';
 import { compileMilkdropPresetSource } from './compiler';
@@ -30,12 +30,25 @@ import { MilkdropOverlay } from './overlay';
 import { consumeRequestedMilkdropOverlayTab } from './overlay-intent';
 import { consumeRequestedMilkdropPresetSelection } from './preset-selection';
 import { createMilkdropRendererAdapter } from './renderer-adapter-factory';
+import { createMilkdropBackendFailover } from './runtime/backend-fallback';
 import { createMilkdropCatalogCoordinator } from './runtime/catalog-coordinator';
+import { buildAgentMilkdropDebugSnapshot } from './runtime/debug-snapshot';
 import { createMilkdropRuntimeInteractionPresenter } from './runtime/interaction-presenter';
+import {
+  applyMilkdropInteractionResponse,
+  buildMilkdropInputSignalOverrides,
+  getMilkdropDetailScale,
+} from './runtime/interaction-response';
+import {
+  buildBlendStateForRender,
+  buildRenderFrameState,
+  shouldAutoAdvancePreset,
+} from './runtime/lifecycle';
 import { createMilkdropPresetFileActions } from './runtime/preset-file-actions';
 import { createMilkdropPresetNavigationController } from './runtime/preset-navigation-controller';
 import { createMilkdropRuntimePreferences } from './runtime/runtime-preferences';
 import { cloneBlendState, estimateFrameBlendWorkload } from './runtime/session';
+import { resolveStartupPresetId } from './runtime/startup';
 import {
   installRequestedOverlayTabListener,
   installRequestedPresetListener,
@@ -44,16 +57,12 @@ import { createMilkdropSignalTracker } from './runtime-signals';
 import type {
   MilkdropCompiledPreset,
   MilkdropFrameState,
-  MilkdropGpuGeometryHints,
-  MilkdropGpuInteractionPayload,
-  MilkdropPostVisual,
   MilkdropRuntimeSignals,
 } from './types';
 import { createMilkdropVM } from './vm';
 import {
   getDisabledMilkdropWebGpuOptimizationFlags,
   resolveMilkdropWebGpuOptimizationFlags,
-  shouldFallbackMilkdropPresetToWebgl,
 } from './webgpu-optimization-flags';
 
 const DEFAULT_PRESET_SOURCE = `title=Signal Bloom
@@ -130,484 +139,6 @@ wave_0_per_point1=y = y + sin(sample * pi * 12 + time) * 0.06
 shape_0_per_frame1=rad = 0.14 + beat_pulse * 0.08
 `;
 
-function sanitizeRuntimeSignals(signals: MilkdropRuntimeSignals) {
-  const {
-    frequencyData: _frequencyData,
-    waveformData: _waveformData,
-    ...rest
-  } = signals;
-  return rest;
-}
-
-const clamp = (value: number, min: number, max: number) =>
-  Math.max(min, Math.min(max, value));
-
-function normalizeSceneTranslation(value: number) {
-  return clamp(value, -0.22, 0.22);
-}
-
-function transformScenePositions(
-  positions: number[],
-  {
-    offsetX,
-    offsetY,
-    rotation,
-    scale,
-  }: { offsetX: number; offsetY: number; rotation: number; scale: number },
-) {
-  if (
-    positions.length === 0 ||
-    (Math.abs(offsetX) < 0.0001 &&
-      Math.abs(offsetY) < 0.0001 &&
-      Math.abs(rotation) < 0.0001 &&
-      Math.abs(scale - 1) < 0.0001)
-  ) {
-    return positions;
-  }
-
-  const cos = Math.cos(rotation);
-  const sin = Math.sin(rotation);
-  const transformed = [...positions];
-
-  for (let index = 0; index < transformed.length; index += 3) {
-    const x = transformed[index] ?? 0;
-    const y = transformed[index + 1] ?? 0;
-    const scaledX = x * scale;
-    const scaledY = y * scale;
-    transformed[index] = scaledX * cos - scaledY * sin + offsetX;
-    transformed[index + 1] = scaledX * sin + scaledY * cos + offsetY;
-  }
-
-  return transformed;
-}
-
-function nudgeCenter(value: number, offset: number) {
-  return clamp(value + offset * 0.5, 0, 1);
-}
-
-function enhancePostEffects(
-  post: MilkdropPostVisual,
-  {
-    offsetX,
-    offsetY,
-    rotation,
-    pinchDelta,
-    dragBoost,
-  }: {
-    offsetX: number;
-    offsetY: number;
-    rotation: number;
-    pinchDelta: number;
-    dragBoost: number;
-  },
-): MilkdropPostVisual {
-  return {
-    ...post,
-    warp: clamp(
-      post.warp + dragBoost * 0.28 + Math.abs(pinchDelta) * 0.12,
-      0,
-      1,
-    ),
-    videoEchoZoom: clamp(post.videoEchoZoom + pinchDelta * 0.12, 0.85, 1.35),
-    shaderControls: {
-      ...post.shaderControls,
-      offsetX: clamp(post.shaderControls.offsetX + offsetX * 0.5, -0.3, 0.3),
-      offsetY: clamp(post.shaderControls.offsetY + offsetY * 0.5, -0.3, 0.3),
-      rotation: clamp(
-        post.shaderControls.rotation + rotation * 0.85,
-        -1.6,
-        1.6,
-      ),
-      zoom: clamp(post.shaderControls.zoom + pinchDelta * 0.3, 0.72, 1.45),
-      warpScale: clamp(
-        post.shaderControls.warpScale + pinchDelta * 0.24 + dragBoost * 0.16,
-        0,
-        2,
-      ),
-    },
-  };
-}
-
-function enhanceGpuGeometry(
-  gpuGeometry: MilkdropGpuGeometryHints,
-  {
-    offsetX,
-    offsetY,
-    rotation,
-    scale,
-    pinchDelta,
-  }: {
-    offsetX: number;
-    offsetY: number;
-    rotation: number;
-    scale: number;
-    pinchDelta: number;
-  },
-): MilkdropGpuGeometryHints {
-  return {
-    ...gpuGeometry,
-    mainWave: gpuGeometry.mainWave
-      ? {
-          ...gpuGeometry.mainWave,
-          centerX: nudgeCenter(gpuGeometry.mainWave.centerX, offsetX),
-          centerY: nudgeCenter(gpuGeometry.mainWave.centerY, -offsetY),
-          scale: clamp(gpuGeometry.mainWave.scale * scale, 0.45, 2.4),
-        }
-      : null,
-    trailWaves: gpuGeometry.trailWaves.map((wave) => ({
-      ...wave,
-      centerX: nudgeCenter(wave.centerX, offsetX),
-      centerY: nudgeCenter(wave.centerY, -offsetY),
-      scale: clamp(wave.scale * scale, 0.45, 2.4),
-    })),
-    customWaves: gpuGeometry.customWaves.map((wave) => ({
-      ...wave,
-      centerX: nudgeCenter(wave.centerX, offsetX),
-      centerY: nudgeCenter(wave.centerY, -offsetY),
-      scaling: clamp(wave.scaling * scale, 0.45, 2.4),
-    })),
-    meshField: gpuGeometry.meshField
-      ? {
-          ...gpuGeometry.meshField,
-          rotation: gpuGeometry.meshField.rotation + rotation * 0.9,
-          zoom: clamp(gpuGeometry.meshField.zoom - pinchDelta * 0.2, 0.5, 2.6),
-          warp: clamp(
-            gpuGeometry.meshField.warp + Math.abs(pinchDelta) * 0.1,
-            0,
-            1.4,
-          ),
-        }
-      : null,
-    motionVectorField: gpuGeometry.motionVectorField
-      ? {
-          ...gpuGeometry.motionVectorField,
-          rotation: gpuGeometry.motionVectorField.rotation + rotation * 0.9,
-          zoom: clamp(
-            gpuGeometry.motionVectorField.zoom - pinchDelta * 0.2,
-            0.5,
-            2.6,
-          ),
-          warp: clamp(
-            gpuGeometry.motionVectorField.warp + Math.abs(pinchDelta) * 0.1,
-            0,
-            1.4,
-          ),
-        }
-      : null,
-  };
-}
-
-export function applyMilkdropInteractionResponse(
-  frameState: MilkdropFrameState,
-  input: UnifiedInputState | null,
-  backend: 'webgl' | 'webgpu' = 'webgl',
-): MilkdropFrameState {
-  if (!input) {
-    return frameState;
-  }
-
-  const gesture = input.gesture;
-  const dragBoost = clamp(input.performance.dragIntensity, 0, 1);
-  const pinchDelta = clamp((gesture?.scale ?? 1) - 1, -0.45, 0.7);
-  const rotation = clamp(gesture?.rotation ?? 0, -Math.PI / 2, Math.PI / 2);
-  const offsetX = normalizeSceneTranslation(
-    input.dragDelta.x * 0.9 + (gesture?.translation.x ?? 0) * 0.45,
-  );
-  const offsetY = normalizeSceneTranslation(
-    input.dragDelta.y * 0.9 + (gesture?.translation.y ?? 0) * 0.45,
-  );
-  const scale = clamp(1 + pinchDelta * 0.45, 0.7, 1.55);
-  const waveAlphaMultiplier = clamp(
-    1 + Math.abs(pinchDelta) * 0.14 + dragBoost * 0.06,
-    0.72,
-    1.35,
-  );
-  const meshAlphaMultiplier = clamp(1 + Math.abs(pinchDelta) * 0.12, 0.72, 1.3);
-  const motionVectorAlphaMultiplier = clamp(
-    1 + Math.abs(pinchDelta) * 0.16 + dragBoost * 0.08,
-    0.72,
-    1.4,
-  );
-  const usesGpuInteractionPayload = backend === 'webgpu';
-
-  if (
-    dragBoost < 0.001 &&
-    Math.abs(pinchDelta) < 0.001 &&
-    Math.abs(rotation) < 0.001 &&
-    Math.abs(offsetX) < 0.001 &&
-    Math.abs(offsetY) < 0.001
-  ) {
-    return frameState;
-  }
-
-  const interaction: MilkdropGpuInteractionPayload | null =
-    usesGpuInteractionPayload
-      ? {
-          waves: {
-            offsetX,
-            offsetY,
-            rotation,
-            scale,
-            alphaMultiplier: waveAlphaMultiplier,
-          },
-          mesh: {
-            offsetX,
-            offsetY,
-            rotation,
-            scale,
-            alphaMultiplier: meshAlphaMultiplier,
-          },
-          motionVectors: {
-            offsetX,
-            offsetY,
-            rotation,
-            scale,
-            alphaMultiplier: motionVectorAlphaMultiplier,
-          },
-        }
-      : null;
-
-  return {
-    ...frameState,
-    interaction,
-    mainWave: {
-      ...frameState.mainWave,
-      positions:
-        usesGpuInteractionPayload && frameState.gpuGeometry.mainWave
-          ? frameState.mainWave.positions
-          : transformScenePositions(frameState.mainWave.positions, {
-              offsetX,
-              offsetY,
-              rotation,
-              scale,
-            }),
-      thickness: clamp(frameState.mainWave.thickness + dragBoost * 0.8, 1, 12),
-    },
-    customWaves: frameState.customWaves.map((wave, index) => ({
-      ...wave,
-      positions:
-        usesGpuInteractionPayload &&
-        Boolean(frameState.gpuGeometry.customWaves[index])
-          ? wave.positions
-          : transformScenePositions(wave.positions, {
-              offsetX,
-              offsetY,
-              rotation,
-              scale,
-            }),
-    })),
-    trails: frameState.trails.map((trail, index) => ({
-      ...trail,
-      positions:
-        usesGpuInteractionPayload &&
-        Boolean(frameState.gpuGeometry.trailWaves[index])
-          ? trail.positions
-          : transformScenePositions(trail.positions, {
-              offsetX,
-              offsetY,
-              rotation,
-              scale,
-            }),
-    })),
-    mesh: {
-      ...frameState.mesh,
-      positions:
-        usesGpuInteractionPayload && frameState.gpuGeometry.meshField
-          ? frameState.mesh.positions
-          : transformScenePositions(frameState.mesh.positions, {
-              offsetX,
-              offsetY,
-              rotation,
-              scale,
-            }),
-      alpha: clamp(frameState.mesh.alpha + Math.abs(pinchDelta) * 0.12, 0, 1),
-    },
-    shapes: frameState.shapes.map((shape) => ({
-      ...shape,
-      x: nudgeCenter(shape.x, offsetX),
-      y: nudgeCenter(shape.y, -offsetY),
-      radius: clamp(shape.radius * scale, 0.02, 0.6),
-      rotation: shape.rotation + rotation,
-    })),
-    motionVectors: frameState.motionVectors.map((vector) => ({
-      ...vector,
-      positions:
-        usesGpuInteractionPayload && frameState.gpuGeometry.motionVectorField
-          ? vector.positions
-          : transformScenePositions(vector.positions, {
-              offsetX,
-              offsetY,
-              rotation,
-              scale,
-            }),
-    })),
-    post: enhancePostEffects(frameState.post, {
-      offsetX,
-      offsetY,
-      rotation,
-      pinchDelta,
-      dragBoost,
-    }),
-    gpuGeometry: enhanceGpuGeometry(frameState.gpuGeometry, {
-      offsetX,
-      offsetY,
-      rotation,
-      scale,
-      pinchDelta,
-    }),
-  };
-}
-
-export function getMilkdropDetailScale({
-  backend,
-  particleScale,
-  particleBudget,
-  shaderQuality = 'balanced',
-}: {
-  backend: 'webgl' | 'webgpu';
-  particleScale?: number;
-  particleBudget: number;
-  shaderQuality?: ShaderQuality;
-}) {
-  const baseScale = (particleScale ?? 1) * particleBudget;
-  const backendBoost = backend === 'webgpu' ? 1.55 : 1.1;
-  const shaderQualityScale =
-    shaderQuality === 'low' ? 0.72 : shaderQuality === 'high' ? 1.2 : 1;
-  return Math.min(
-    2,
-    Math.max(0.5, baseScale * backendBoost * shaderQualityScale),
-  );
-}
-
-export function buildMilkdropInputSignalOverrides(
-  input: UnifiedInputState | null,
-  target: Partial<MilkdropRuntimeSignals> = {},
-): Partial<MilkdropRuntimeSignals> {
-  const gesture = input?.gesture;
-  const performance = input?.performance;
-  const sourceFlags = performance?.sourceFlags;
-  const actions = performance?.actions;
-  const inputSpeed = Math.hypot(
-    input?.dragDelta.x ?? 0,
-    input?.dragDelta.y ?? 0,
-  );
-
-  return Object.assign(target, {
-    inputX: input?.normalizedCentroid.x ?? 0,
-    inputY: input?.normalizedCentroid.y ?? 0,
-    input_x: input?.normalizedCentroid.x ?? 0,
-    input_y: input?.normalizedCentroid.y ?? 0,
-    inputDx: input?.dragDelta.x ?? 0,
-    inputDy: input?.dragDelta.y ?? 0,
-    input_dx: input?.dragDelta.x ?? 0,
-    input_dy: input?.dragDelta.y ?? 0,
-    inputSpeed: inputSpeed,
-    input_speed: inputSpeed,
-    inputPressed: input?.isPressed ? 1 : 0,
-    input_pressed: input?.isPressed ? 1 : 0,
-    inputJustPressed: input?.justPressed ? 1 : 0,
-    input_just_pressed: input?.justPressed ? 1 : 0,
-    inputJustReleased: input?.justReleased ? 1 : 0,
-    input_just_released: input?.justReleased ? 1 : 0,
-    inputCount: input?.pointerCount ?? 0,
-    input_count: input?.pointerCount ?? 0,
-    gestureScale: gesture?.scale ?? 1,
-    gesture_scale: gesture?.scale ?? 1,
-    gestureRotation: gesture?.rotation ?? 0,
-    gesture_rotation: gesture?.rotation ?? 0,
-    gestureTranslateX: gesture?.translation.x ?? 0,
-    gestureTranslateY: gesture?.translation.y ?? 0,
-    gesture_translate_x: gesture?.translation.x ?? 0,
-    gesture_translate_y: gesture?.translation.y ?? 0,
-    hoverActive: performance?.hoverActive ? 1 : 0,
-    hover_active: performance?.hoverActive ? 1 : 0,
-    hoverX: performance?.hover?.x ?? 0,
-    hoverY: performance?.hover?.y ?? 0,
-    hover_x: performance?.hover?.x ?? 0,
-    hover_y: performance?.hover?.y ?? 0,
-    wheelDelta: performance?.wheelDelta ?? 0,
-    wheel_delta: performance?.wheelDelta ?? 0,
-    wheelAccum: performance?.wheelAccum ?? 0,
-    wheel_accum: performance?.wheelAccum ?? 0,
-    dragIntensity: performance?.dragIntensity ?? 0,
-    drag_intensity: performance?.dragIntensity ?? 0,
-    dragAngle: performance?.dragAngle ?? 0,
-    drag_angle: performance?.dragAngle ?? 0,
-    accentPulse: performance?.accentPulse ?? 0,
-    accent_pulse: performance?.accentPulse ?? 0,
-    actionAccent: actions?.accent ?? 0,
-    action_accent: actions?.accent ?? 0,
-    actionModeNext: actions?.modeNext ?? 0,
-    action_mode_next: actions?.modeNext ?? 0,
-    actionModePrevious: actions?.modePrevious ?? 0,
-    action_mode_previous: actions?.modePrevious ?? 0,
-    actionPresetNext: actions?.presetNext ?? 0,
-    action_preset_next: actions?.presetNext ?? 0,
-    actionPresetPrevious: actions?.presetPrevious ?? 0,
-    action_preset_previous: actions?.presetPrevious ?? 0,
-    actionQuickLook1: actions?.quickLook1 ?? 0,
-    action_quick_look_1: actions?.quickLook1 ?? 0,
-    actionQuickLook2: actions?.quickLook2 ?? 0,
-    action_quick_look_2: actions?.quickLook2 ?? 0,
-    actionQuickLook3: actions?.quickLook3 ?? 0,
-    action_quick_look_3: actions?.quickLook3 ?? 0,
-    actionRemix: actions?.remix ?? 0,
-    action_remix: actions?.remix ?? 0,
-    inputSourcePointer: sourceFlags?.pointer ? 1 : 0,
-    input_source_pointer: sourceFlags?.pointer ? 1 : 0,
-    inputSourceKeyboard: sourceFlags?.keyboard ? 1 : 0,
-    input_source_keyboard: sourceFlags?.keyboard ? 1 : 0,
-    inputSourceGamepad: sourceFlags?.gamepad ? 1 : 0,
-    input_source_gamepad: sourceFlags?.gamepad ? 1 : 0,
-    inputSourceMouse: sourceFlags?.mouse ? 1 : 0,
-    input_source_mouse: sourceFlags?.mouse ? 1 : 0,
-    inputSourceTouch: sourceFlags?.touch ? 1 : 0,
-    input_source_touch: sourceFlags?.touch ? 1 : 0,
-    inputSourcePen: sourceFlags?.pen ? 1 : 0,
-    input_source_pen: sourceFlags?.pen ? 1 : 0,
-  });
-}
-
-function buildAgentMilkdropDebugSnapshot({
-  activePresetId,
-  compiledPreset,
-  frameState,
-  status,
-  adaptiveQuality,
-}: {
-  activePresetId: string | null;
-  compiledPreset: MilkdropCompiledPreset | null;
-  frameState: MilkdropFrameState | null;
-  status: string | null;
-  adaptiveQuality?: AdaptiveQualityState | null;
-}) {
-  if (!frameState) {
-    return {
-      activePresetId,
-      status,
-      adaptiveQuality,
-      frameState: null,
-      title: compiledPreset?.title ?? null,
-    };
-  }
-
-  return {
-    activePresetId,
-    status,
-    adaptiveQuality,
-    title: compiledPreset?.title ?? frameState.title,
-    frameState: {
-      presetId: frameState.presetId,
-      title: frameState.title,
-      signals: sanitizeRuntimeSignals(frameState.signals),
-      variables: frameState.variables,
-      mainWave: frameState.mainWave,
-      shapes: frameState.shapes,
-      post: frameState.post,
-    },
-  };
-}
-
 export function createMilkdropExperience({
   container,
   quality,
@@ -653,7 +184,6 @@ export function createMilkdropExperience({
   );
   let transitionMode = preferences.getTransitionMode();
   let lastPresetSwitchAt = performance.now();
-  let fallbackTriggered = false;
   let lastStatusMessage: string | null = null;
   let disposeKeyboardShortcuts: (() => void) | null = null;
   let disposeRequestedPresetListener: (() => void) | null = null;
@@ -667,6 +197,13 @@ export function createMilkdropExperience({
     shaderEnabled: false,
     videoEchoEnabled: false,
   };
+
+  const backendFailover = createMilkdropBackendFailover({
+    preferences,
+    reload: () => {
+      window.location.reload();
+    },
+  });
 
   const updateAgentDebugSnapshot = () => {
     if (!isAgentMode()) {
@@ -720,11 +257,10 @@ export function createMilkdropExperience({
   });
 
   const shouldFallbackToWebgl = (compiled: MilkdropCompiledPreset) =>
-    shouldFallbackMilkdropPresetToWebgl({
-      backend: activeBackend,
-      compatibilityMode: isCompatibilityModeEnabled(),
-      descriptorPlan: compiled.ir.compatibility.gpuDescriptorPlans.webgpu,
-      flags: webgpuOptimizationFlags,
+    backendFailover.shouldFallback({
+      compiled,
+      activeBackend,
+      webgpuOptimizationFlags,
     });
 
   const triggerWebglFallback = ({
@@ -734,13 +270,11 @@ export function createMilkdropExperience({
     presetId: string;
     reason: string;
   }) => {
-    if (fallbackTriggered || activeBackend !== 'webgpu') {
-      return;
-    }
-    fallbackTriggered = true;
-    preferences.recordFallback({ presetId, reason });
-    setCompatibilityMode(true);
-    window.location.reload();
+    backendFailover.trigger({
+      presetId,
+      reason,
+      activeBackend,
+    });
   };
 
   const navigation = createMilkdropPresetNavigationController({
@@ -998,16 +532,17 @@ export function createMilkdropExperience({
       if (collectionEntry && requestedCollectionTag) {
         overlay.setActiveCollectionTag(requestedCollectionTag);
       }
-      const requestedPresetId = consumeRequestedMilkdropPresetSelection();
-      const preferredStartupPresetId =
-        requestedPresetId ??
-        preferences.getStartupPresetId(initialPresetId) ??
-        collectionEntry?.id;
-      const startupPresetId =
-        preferredStartupPresetId &&
-        navigation.isBackendSelectable(preferredStartupPresetId, activeBackend)
-          ? preferredStartupPresetId
-          : navigation.getFirstSelectablePresetId(activeBackend);
+      const requestedPresetId =
+        consumeRequestedMilkdropPresetSelection() ?? null;
+      const startupPresetId = resolveStartupPresetId({
+        requestedPresetId,
+        preferredStartupPresetId:
+          preferences.getStartupPresetId(initialPresetId) ?? null,
+        collectionEntryId: collectionEntry?.id ?? null,
+        isBackendSelectable: navigation.isBackendSelectable,
+        getFirstSelectablePresetId: navigation.getFirstSelectablePresetId,
+        activeBackend,
+      });
       if (startupPresetId) {
         await navigation.selectPreset(startupPresetId, {
           recordHistory: false,
@@ -1154,9 +689,13 @@ export function createMilkdropExperience({
       );
 
       if (
-        autoplay &&
-        catalogCoordinator.getCatalogEntries().length > 1 &&
-        now - lastPresetSwitchAt > Math.max(12000, blendDuration * 1000 + 6000)
+        shouldAutoAdvancePreset({
+          autoplay,
+          catalogSize: catalogCoordinator.getCatalogEntries().length,
+          now,
+          lastPresetSwitchAt,
+          blendDuration,
+        })
       ) {
         void navigation.selectRandomPreset();
       }
@@ -1169,37 +708,21 @@ export function createMilkdropExperience({
       updateAgentDebugSnapshot();
       const canBlendCurrentFrame =
         estimateFrameBlendWorkload(currentFrameState) < 900;
-      const activeBlendState =
-        transitionMode === 'blend' &&
-        frame.performance.shaderQuality !== 'low' &&
-        canBlendCurrentFrame &&
-        blendState &&
-        now < blendEndAtMs
-          ? {
-              ...blendState,
-              alpha:
-                1 -
-                (now - (blendEndAtMs - blendDuration * 1000)) /
-                  (blendDuration * 1000),
-            }
-          : null;
+      const activeBlendState = buildBlendStateForRender({
+        transitionMode,
+        shaderQuality: frame.performance.shaderQuality,
+        canBlendCurrentFrame,
+        blendState,
+        now,
+        blendEndAtMs,
+        blendDuration,
+      });
 
-      const renderFrameState =
-        frame.performance.shaderQuality === 'low' &&
-        (currentFrameState.post.shaderEnabled ||
-          currentFrameState.post.videoEchoEnabled)
-          ? {
-              ...currentFrameState,
-              post: Object.assign(
-                lowQualityPostOverride,
-                currentFrameState.post,
-                {
-                  shaderEnabled: false,
-                  videoEchoEnabled: false,
-                },
-              ),
-            }
-          : currentFrameState;
+      const renderFrameState = buildRenderFrameState({
+        frameState: currentFrameState,
+        shaderQuality: frame.performance.shaderQuality,
+        lowQualityPostOverride,
+      });
 
       const renderStartAt = performance.now();
       const adapterPresentedFrame = adapter.render({
